+++ NEWS.md
@@ -1,3 +1,8 @@
+# BiocChallenges 0.0.4
+
+* Query the datetime of the latest push to each contributing repository.
+* Cache queries to the GitHub API for individual repositories.
+
 # BiocChallenges 0.0.3
 
 * Move challenge base Rmd file to `inst/` for reuse in downstream packages.